+++ Changelog.md
@@ -1,7 +1,7 @@
 # Factureaza PHP SDK Changelog
 
-## Unreleased
-##### 2022-XX-YY
+## 1.1.0
+##### 2022-10-11
 
 - Added lower and upper annotation support to Invoices
 - Fixed `createInvoice::withClient` method when using with a Client model